tests/wpt/tests/css/css-borders/tentative/parsing/border-shape-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'border-shape' with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#border-shape">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_invalid_value("border-shape", "circle() none");
test_invalid_value("border-shape", "polygon(bla 10px 10px, 100px 10px, 10px 100px)");
test_invalid_value("border-shape", "shape()");
test_invalid_value("border-shape", "nonsense");
test_invalid_value("border-shape", "1px");

</script>
